% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_add.R
\name{add_stat}
\alias{add_stat}
\title{Add a Statistic to \code{regg} Object}
\usage{
add_stat(x, stat, ...)
}
\arguments{
\item{x}{object of class \code{regg} or \code{rgo_model}}

\item{stat}{function for estimating the statistic, to be applied to the regression model}

\item{...}{objects passed to methods}
}
\value{
the object \code{x}
}
\description{
This function is used to add a statistic function to a \code{ggr} object. It
is intended to be called by the \code{regg} method of the original regression
statistic function.
}
