% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_add.R
\name{add_test}
\alias{add_test}
\title{Add a Test to \code{regg} Object}
\usage{
add_test(x, test, ...)
}
\arguments{
\item{x}{object of class \code{regg} or \code{rgo_model}}

\item{test}{function for estimating the test, to be applied to the regression model}

\item{...}{objects passed to methods}

\item{args}{a quoted list of function arguments to be passed instead of \code{...}}
}
\value{
the object \code{x}
}
\description{
This function is used to add a test function to a \code{ggr} object. It
is intended to be called by the \code{regg} method of the original regression
test function.
}
