% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_x.R
\name{add_x}
\alias{add_x}
\title{Estimate a New Regression by Adding Independent Variables}
\usage{
add_x(x, ...)
}
\arguments{
\item{x}{\code{rgo} object containing a regression model.}

\item{...}{If unnamed, variables to be added as independent variables in the
new regression using \code{\link{reg_select}}. If named, additional fields
passed to \code{\link{regg}}.}
}
\value{
environment of class \code{regg}
}
\description{
Builds a new regression model using the regression in \code{x} as the basic template,
including the model terms. Unnamed objects passed to \code{add_x} are included
as additional independent variables, using grammar from \code{\link{reg_select}}.
Variables from the previous regression can be removed with the \code{-} prefix.
}
