% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_helpers.R
\name{as_stat}
\alias{as_stat}
\title{Create an \code{rgo_test} Object}
\usage{
as_stat(x, label = NULL, ...)
}
\arguments{
\item{label}{character name to display when printed}

\item{...}{Other objects included in the attributes.}

\item{stat, df, pval}{numeric values or vectors}
}
\value{
list of class \code{"rgo_test"}
}
\description{
\code{rgo_test} objects are lists containing statistics, associated degrees
of freedom, p-values, and other fields containing relevant test information.
}
