% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_helpers.R
\name{as_test}
\alias{as_test}
\title{Create an \code{rgo_test} Object}
\usage{
as_test(stat, pval = NULL, df = NULL, label = NULL, ...)
}
\arguments{
\item{stat, df, pval}{numeric values or vectors}

\item{label}{character name to display when printed}

\item{...}{Other objects include in the list.}
}
\value{
list of class \code{"rgo_test"}
}
\description{
\code{rgo_test} objects are lists containing statistics, associated degrees
of freedom, p-values, and other fields containing relevant test information.
}
