% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explainEGG.R
\name{explainEGG}
\alias{explainEGG}
\title{Explaining the EGG metadata categories}
\usage{
explainEGG()
}
\value{
Data frame containing EGG metdata category names and descriptions
}
\description{
Returns a data frame with each metadata category and a brief description.
For more details about the data set, please refer to the EGG manuscript
at https://doi.org/10.1186/s42523-025-00493-x.
}
\examples{

explainEGG()
}
