% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract}
\alias{extract}
\alias{$.rgo_model}
\alias{[[.rgo_model}
\title{Extract Elements from Regression Models}
\usage{
\method{$}{rgo_model}(x, name)

\method{[[}{rgo_model}(x, i, inherits = FALSE, ...)
}
\arguments{
\item{x}{object of class \code{rgo}}

\item{name, i}{character describing object to extract}

\item{inherits}{should earlier models be searched? Defaults to \code{TRUE}
when using \code{`$`}, and \code{FALSE} when using \code{`[[`}.}

\item{...}{objects passed to methods}
}
\value{
R object
}
\description{
Extract Elements from Regression Models
}
