% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reg.R
\name{extract_reg}
\alias{extract_reg}
\alias{coef.rgo_model}
\alias{fitted.rgo_model}
\alias{residuals.rgo_model}
\alias{weights.rgo_model}
\title{Extract Regression Objects}
\usage{
\method{coef}{rgo_model}(object, write = TRUE, ...)

\method{fitted}{rgo_model}(object, write = TRUE, ...)

\method{residuals}{rgo_model}(
  object,
  type = c("working", "response", "deviance", "pearson", "partial"),
  write = TRUE,
  ...
)

\method{weights}{rgo_model}(object, write = TRUE, ...)
}
\arguments{
\item{write}{should the requested object be inserted in \code{x}, if it not there}

\item{...}{objects passed to methods}

\item{x}{\code{regg} or \code{rgo_model} object}
}
\value{
the object, extracted from \code{x}
}
\description{
Obtain common regression objects such as the coefficients, fitted values, and residuals.
}
