% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devel_helpers.R
\name{find_else}
\alias{find_else}
\alias{fit_find}
\title{Find Object in Regression Model Else Evaluate Expression}
\usage{
find_else(x, what, expr, mode = "any", inherits = TRUE, write = TRUE)

fit_find(x, what, expr, mode = "any", inherits = FALSE, write = TRUE, ...)
}
\arguments{
\item{x}{object of class \code{rgo} to be searched}

\item{what}{character vector to search}

\item{expr}{expression to be evaluated if object is not found}

\item{mode}{the mode or type of object sought. See details in \code{\link[base]{exists}}.}

\item{inherits}{should previous models be searched?}

\item{write}{should the evaluated expression be inserted in \code{x} as \code{what}?}

\item{...}{objects passed to \code{\link{fit.rgo_model}}}
}
\value{
R object
}
\description{
These functions look for an object within a \code{rgo} object. If not found,
an expression, \code{expr}, will be evaluated and returned. The result
may be written to the object so that it is easily available in the future.
\code{fit_find} attempts to fit a regression model first - if it is unfitted -
then searches for the object. These functions are intended for \code{regg}
developers.
}
