% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devel_helpers.R
\name{find_models}
\alias{find_models}
\title{Return All Nested Models in \code{rgo} Regression Object}
\usage{
find_models(x)
}
\arguments{
\item{x}{object of class \code{rgo_m} to be searched}
}
\value{
list containing objects containing class \code{rgo_m}
}
\description{
Return All Nested Models in \code{rgo} Regression Object
}
