% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\alias{fit.regg}
\alias{fit.rgo_model}
\alias{refit.regg}
\alias{refit.rgo_model}
\title{Fit a \code{regg} Regression Model}
\usage{
\method{fit}{regg}(x, ...)

\method{fit}{rgo_model}(x, refit = FALSE, keep = FALSE, ...)

\method{refit}{regg}(x, ...)

\method{refit}{rgo_model}(x, ...)
}
\arguments{
\item{x}{object to be fitted}

\item{...}{objects passed to methods}

\item{refit}{should the model be fitted again if it is already fitted?}

\item{keep}{should the \code{X} and \code{y} matrices be kept after fitting?}
}
\value{
\code{ggr_model} or \code{ggreg} object
}
\description{
Fit a \code{regg} Regression Model
}
\details{
\code{\link{regg}} and \code{rgo_model} objects are initialized with a flag \code{"is_fitted"}
set to \code{FALSE}. \code{ggr_fit} first checks whether this flag is \code{TRUE}.
If so, the object is returned, unless the \code{refit} argument is set to \code{TRUE}.
If the flag is \code{FALSE}, then the following takes place.

First, \code{\link{get_model_matrix}} is run to set the \code{X} and \code{y}
variable fields inside of the environment. Next, the method of fitting the
model is called on the \code{ggr_model} object and the \code{"is_fitted"} flag
is set to \code{TRUE}. Any post-fit functions are applied to model, the standard
errors and other regression statistics are calculated, then statistical tests
are performed.
}
