% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ols.R
\name{fit_ols}
\alias{fit_ols}
\title{Fit a Model using Ordinary Least Squares}
\usage{
fit_ols(model, X, y, weights = NULL, offset = NULL, intercept = NULL)
}
\arguments{
\item{model}{object of class \code{rgo_model}}

\item{X}{numeric matrix of independent variables}

\item{y}{numeric vector or matrix of predictor variables}

\item{weights}{numeric vector of weights or \code{NULL}}

\item{offset}{numeric vector to specify a priori known component to be included in the predictor.}

\item{intercept}{does the model include an intercept? If \code{NULL}, \code{fit_ols} will attempt to determine automatically.}
}
\value{
the regression model
}
\description{
This is used by \code{\link{ols}} and related methods. It is not intended to
be used directly for analysis, but within a regression method to compute the
relevant QR matrix and coefficients, and add the appropriate fields to the
model object. It is a wrapper for \code{\link[stats]{lm.fit}} and
\code{\link[stats]{lm.wfit}}.
}
