% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Find Regression Data}
\usage{
get_data(x, ...)
}
\arguments{
\item{x}{object of class \code{rgo} to be searched}

\item{...}{objects passed to methods}
}
\value{
data.frame or related object
}
\description{
These functions look for a data set within an \code{rgo} object. If used on an
\code{rgo_model}, the underlying data set - subset if specified - will be
returned.
}
