% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reg_stats.R
\name{get_stats}
\alias{get_stats}
\alias{get_mss}
\alias{get_rank}
\alias{get_rss}
\alias{get_se}
\title{Obtain & Extract Regression Statistics}
\usage{
get_mss(x, write = TRUE)

get_rank(x, write = TRUE)

get_rss(x, write = TRUE)

get_se(x, write = TRUE)
}
\arguments{
\item{x}{\code{rgo_model} object}

\item{write}{should the requested object be inserted in \code{x}, if it not found}
}
\value{
the requested object, usually numeric vector or matrix
}
\description{
These functions are intended to be used inside of methods to extract regression
statistics - such as the residual sum of squares (rss) - for calculating other
statistics or tests.
}
\section{Functions}{
\itemize{
\item \code{get_mss()}: Obtain the mean sum of squares

\item \code{get_rank()}: Obtain the rank of the regression

\item \code{get_rss()}: Obtain the sum of squared residuals

\item \code{get_se()}: Obtain the standard errors for the model coefficients

}}
