% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{getEGG}
\alias{getEGG}
\title{Pull EGG Data Files}
\usage{
getEGG(version = "1.0.0", data_type = c("metadata", "table", "taxonomy"))
}
\arguments{
\item{version}{Version of EGG database; default = 1.0.0}
}
\description{
Pull EGG Data Files
}
