% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEGG_version.R
\name{getEGG_version}
\alias{getEGG_version}
\title{Get EGG Version History}
\usage{
getEGG_version()
}
\value{
Data frame with version history and corresponding Zenodo ID.
}
\description{
Get EGG Version History
}
\examples{
getEGG_version()
}
