% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{is_regg}
\alias{is_regg}
\title{Check Whether Input is a 'regg' Object}
\usage{
is_regg(x)
}
\arguments{
\item{x}{R object to be tested}
}
\value{
\code{TRUE} if \code{x} inherits \code{"regg"}, otherwise \code{FALSE}
}
\description{
Check Whether Input is a 'regg' Object
}
