% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{is_rgo_model}
\alias{is_rgo_model}
\title{Check Whether Input is an 'rgo_model' Object}
\usage{
is_rgo_model(x)
}
\arguments{
\item{x}{R object to be tested}
}
\value{
\code{TRUE} if \code{x} inherits \code{"rgo_model"}, otherwise \code{FALSE}
}
\description{
Check Whether Input is an 'rgo_model' Object
}
