% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_helpers.R
\name{is_stat}
\alias{is_stat}
\title{Test Whether Object is an \code{rgo_stat}}
\usage{
is_stat(x)
}
\arguments{
\item{x}{object to be tested}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Test Whether Object is an \code{rgo_stat}
}
