% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_helpers.R
\name{is_test}
\alias{is_test}
\title{Test Whether Object is an \code{rgo_test}}
\usage{
is_test(x)
}
\arguments{
\item{x}{object to be tested}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Test Whether Object is an \code{rgo_test}
}
