% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devel_helpers.R
\name{length_models}
\alias{length_models}
\title{Calculate the Number of Models in an \code{rgo} Object}
\usage{
length_models(x)
}
\arguments{
\item{x}{object of class \code{rgo} to be searched}
}
\value{
R object
}
\description{
Calculate the Number of Models in an \code{rgo} Object
}
