% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_ols.R
\name{ols}
\alias{ols}
\title{Ordinary Least Squares Regression}
\usage{
ols(x, ...)
}
\arguments{
\item{x}{a data set, \code{rgo_model_matrix}, or \code{\link{regg}} object}

\item{...}{If unnamed, the description of the model terms using \code{\link{reg_select}}.
If named, arguments passed to \code{\link{regg}}. See 'details' below.}
}
\value{
object of class 'regg'
}
\description{
Fit a linear model using the \code{regg} ecosystem; equivalent to \code{\link[stats]{lm}}.
}
