% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devel_helpers.R
\name{reg_eval}
\alias{reg_eval}
\title{Evaluate a Expression in \code{regg}}
\usage{
reg_eval(x, where, ...)
}
\arguments{
\item{x}{expression to be evaluated}

\item{where}{\code{rgo} object}

\item{...}{objects passed to \code{\link{reg_select}}}
}
\value{
R object
}
\description{
This function is intended for developers to evaluate quoted expressions passed
to \code{\link{regg}} using \code{\link{reg_select}} semantics.
}
