% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_step.R
\name{reg_step}
\alias{reg_step}
\title{Sequentially Add X Variables to Regression}
\usage{
reg_step(x, ...)
}
\arguments{
\item{x}{\code{rgo} object containing a regression model.}

\item{...}{If unnamed, variables to be added as independent variables in the
new regression using \code{\link{reg_select}}. If named, additional fields
passed to \code{\link{regg}}.}
}
\value{
environment of class \code{regg}
}
\description{
Adds a series of regression models to \code{x}, by sequentially adding the
unnamed objects passed to the function.
}
