% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rEGG-package.R
\docType{package}
\name{rEGG-package}
\alias{rEGG}
\alias{rEGG-package}
\title{rEGG: Equine Gut Group Database}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package provides a user-friendly method to download and interact with the Equine Gut Group (EGG) Database.
}
\author{
\strong{Maintainer}: McAdams Zach \email{zlmg2b@umsystem.edu}

}
\keyword{internal}
