% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se_default.R
\name{se_default}
\alias{se_default}
\title{Calculate Default Standard Errors}
\usage{
se_default(x, ...)
}
\arguments{
\item{x}{object of class \code{rgo_model} or \code{regg} from which standard errors are calculated.}

\item{...}{Other arguments passed on to methods}
}
\value{
environment of class \code{regg} or \code{rgo_model}
}
\description{
Default function for estimating standard errors, as calculated from the QR
matrix, if available.
}
