% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_common.R
\name{stats_common}
\alias{stats_common}
\alias{stat_adj_r2}
\alias{stat_conf}
\alias{stat_mae}
\alias{stat_n}
\alias{stat_r2}
\alias{stat_rmse}
\title{Calculate and Add Common Regression Statistics}
\usage{
stat_adj_r2(x, ...)

stat_conf(x, ...)

stat_mae(x, ...)

stat_n(x, ...)

stat_r2(x, ...)

stat_rmse(x, ...)
}
\arguments{
\item{x}{\code{regg} or \code{rgo_model} object}

\item{...}{objects passed to methods}
}
\value{
numeric value of statistic
}
\description{
Common statistics include the number of observations \code{(N)}, the R-squared
\code{(R2)} and adjusted R-squared \code{(adj_R2)}. Each are included in the
\code{"stats"} field of each \code{rgo_model}.
}
