% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetEGG.R
\name{subsetEGG}
\alias{subsetEGG}
\title{Subsetting EGG Files}
\usage{
subsetEGG(
  samples = NULL,
  metadata = NULL,
  table = NULL,
  taxonomy = NULL,
  output_file = NULL
)
}
\arguments{
\item{samples}{Object containing sample IDs to be subset from EGG dataset}

\item{metadata}{EGG metadata object}

\item{table}{EGG table object}

\item{taxonomy}{EGG taxonomy object}

\item{output_file}{Name of output directory to be created}
}
\value{
Writes three comma-separated values files
}
\description{
Subsetting EGG Files
}
\examples{

metadata = data.frame(
 sampleid = c("A", "B"),
 group = c(1, 2)
)

samples = c("A", "B")

table = data.frame(
 featureid = c("a1", "b1"),
 A = c(1,0),
 B = c(0,1)
)

taxonomy = data.frame(
 featureid = c("a1", "b1"),
 taxon = c("Bacteria.Bacillota", "Bacteria.Pseudomonadota")
)

subsetEGG(samples = samples,
          metadata = metadata,
          table = table,
          taxonomy = taxonomy,
          output_file = 'filtered_data'
)

unlink('filtered_data', recursive = TRUE)

}
