% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_f.R
\name{test_f}
\alias{test_f}
\title{Perform F-Test on Regression}
\usage{
test_f(x, ...)
}
\arguments{
\item{x}{object of class \code{rgo_model} or \code{regg}.}

\item{...}{Other arguments passed on to methods}
}
\value{
environment of class \code{regg} or \code{rgo_model}
}
\description{
Estimates the F-statistic and p-values associated with the hypothesis test
that all coefficients are equal to 0, unless otherwise specified.
}
