% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_t.R
\name{test_t}
\alias{test_t}
\title{Perform t-Test on Regression Coefficients}
\usage{
test_t(x, ...)
}
\arguments{
\item{x}{object of class \code{rgo_model} or \code{regg} from which standard errors are calculated.}

\item{...}{Other arguments passed on to methods}
}
\value{
environment of class \code{regg} or \code{rgo_model}
}
\description{
Estimates the t-statistic and p-values associated with the hypothesis test
that each coefficient is equal to 0.
}
