% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy}
\alias{tidy}
\alias{augment.rgo_m}
\alias{components.rgo_model}
\alias{components.regg}
\alias{glance.rgo_model}
\alias{glance.regg}
\alias{tidy.rgo_model}
\alias{tidy.regg}
\title{Methods for Augmenting Data}
\usage{
\method{augment}{rgo_m}(x, ...)

\method{components}{rgo_model}(
  x,
  label = NULL,
  bind = TRUE,
  objects = c("resids", "fitted", "hat", "cooks"),
  ...
)

\method{components}{regg}(x, bind = TRUE, ...)

\method{glance}{rgo_model}(x, label = NULL, bind = TRUE, ...)

\method{glance}{regg}(x, bind = TRUE, ...)

\method{tidy}{rgo_model}(
  x,
  label = NULL,
  bind = TRUE,
  type = c("coef", "test", "stat", "full"),
  add_type = FALSE,
  ...
)

\method{tidy}{regg}(x, bind = TRUE, ...)
}
\arguments{
\item{x}{object of class \code{regg} or \code{rgo_model}}

\item{...}{objects passed to methods}

\item{label}{default label of the model when multiple models are present.}

\item{bind}{should the results be bound into a single tibble data.frame?}

\item{type}{character vector describing the type of results to be returned -
coefficient table or tests.}
}
\value{
\code{tibble} object
}
\description{
These methods extract information about the regression object and place into a
\code{\link[tibble]{tibble}}. \code{tidy} extracts the regression coefficient
table, along with model names if appropriate. Each column \code{glance} represents
a regression statistic such as the R-squared. \code{augment} adds information
related to residuals, fitted values, and so forth to the data set.
}
