% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stars.R
\name{use_stars}
\alias{use_stars}
\alias{stars_default}
\title{Use Specified Significance Indicators}
\usage{
use_stars(x, ..., list = NULL)

stars_default()
}
\arguments{
\item{x}{object of class \code{rgo}}

\item{...}{numeric values representing the noted significance levels, named with the desired mark.}

\item{list}{named list or numeric vector. Used as an alternative to \code{...}.
Alternatively, a function that accepts a vector of numbers and returns a
character vector.}
}
\value{
the object \code{x}
}
\description{
Specify the noted significance levels in regression output and how they are
marked.
}
\section{Functions}{
\itemize{
\item \code{stars_default()}: the default significance levels and markings in R, to be added to \code{list}

}}
