% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_select_window.R
\name{window_functions}
\alias{window_functions}
\alias{lag}
\alias{lead}
\title{Window Functions for \code{reg_select}}
\usage{
lag(x, n = 1L, ...)

lead(x, n = 1L, ...)
}
\arguments{
\item{x}{R object}

\item{n}{integer describing the number of observations to lead or lag}

\item{...}{objects passed to methods}
}
\value{
object with same type as \code{x}
}
\description{
A collection of functions to be used in \code{\link{reg_select}}. The default,
if used outside of \code{reg_select}, is the same as their \link{dplyr} equivalent.
}
