% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEGG.R
\name{getEGG}
\alias{getEGG}
\title{getEGG}
\usage{
getEGG(
  version = "1.1.0",
  data_type = c("metadata", "table", "taxonomy", "biosample")
)
}
\arguments{
\item{version}{Version of EGG database to be downloaded}

\item{data_type}{Type of data to be downloaded (metadata, table, or taxonomic assignments)}
}
\value{
A R object of the requested data type from the EGG database.
}
\description{
getEGG
}
\examples{
metadata = getEGG(version = "1.1.0", data_type = "metadata")

taxonomy = getEGG(version = "1.1.0", data_type = "taxonomy")
}
